+++ src/gradescopeapi/__main__.py
@@ -1,3 +1,6 @@
+# Start fastapi server
+
+
 def main():
     pass
