+++ src/gradescopeapi/classes/_helpers/_assignment_helpers.py
@@ -2,7 +2,7 @@ import requests
 import json
 from datetime import datetime
 
-from gradescopeapi.classes.assignments import Assignment
+from gradescopeapi.classes._data_model import Assignment
 
 
 def check_page_auth(session, endpoint):
