+++ src/gradescopeapi/classes/connection.py
@@ -5,25 +5,28 @@ from gradescopeapi.classes._helpers._login_helpers import (
     login_set_session_cookies,
 )
 
-from gradescopeapi.classes.account import Account
 
+def login(email: str, password: str) -> requests.Session:
+    """Logs into Gradescope and returns the session.
 
-class GSConnection:
-    def __init__(self):
-        self.session = requests.Session()
-        self.logged_in = False
-        self.account = None
+    Args:
+        email (str): The email of the user.
+        password (str): The password of the user.
 
-    def login(self, email, password):
-        # go to homepage to parse hidden authenticity token and to set initial "_gradescope_session" cookie
-        auth_token = get_auth_token_init_gradescope_session(self.session)
+    Returns:
+        requests.Session: The session object.
 
-        # login and set cookies in session. Result bool on whether login was success
-        login_success = login_set_session_cookies(
-            self.session, email, password, auth_token
-        )
-        if login_success:
-            self.logged_in = True
-            self.account = Account(self.session)
-        else:
-            raise ValueError("Invalid credentials.")
+    Raises:
+        ValueError: If the login credentials are invalid.
+    """
+    session = requests.Session()
+
+    # go to homepage to parse hidden authenticity token and to set initial "_gradescope_session" cookie
+    auth_token = get_auth_token_init_gradescope_session(session)
+
+    # login and set cookies in session. Result bool on whether login was success
+    login_success = login_set_session_cookies(session, email, password, auth_token)
+    if not login_success:
+        raise ValueError("Invalid credentials.")
+
+    return session
